+++ sevenn/main/sevenn.py
@@ -46,12 +46,19 @@ def run(args):
     distributed = args.distributed
     distributed_backend = args.distributed_backend
     use_cue = args.enable_cueq
+    use_flash = args.enable_flashTP
 
     if use_cue:
         import sevenn.nn.cue_helper
 
         if not sevenn.nn.cue_helper.is_cue_available():
             raise ImportError('cuEquivariance not installed.')
+
+    if use_flash:
+        import sevenn.nn.flash_helper
+
+        if not sevenn.nn.flash_helper.is_flash_available():
+            raise ImportError('FlashTP not installed or no GPU found.')
 
     if working_dir is None:
         working_dir = os.getcwd()
@@ -108,6 +115,9 @@ def run(args):
             else:
                 model_config[KEY.CUEQUIVARIANCE_CONFIG].update({'use': True})
 
+        if use_flash:
+            model_config[KEY.USE_FLASH_TP] = True
+
         logger.print_config(model_config, data_config, train_config)
         # don't have to distinguish configs inside program
         global_config.update(model_config)
@@ -144,7 +154,13 @@ def cmd_parser_train(parser):
     ag.add_argument(
         '-cueq',
         '--enable_cueq',
-        help='(Not stable!) use cuEquivariance for training',
+        help='use cuEq accelerations for training',
+        action='store_true',
+    )
+    ag.add_argument(
+        '-flashTP',
+        '--enable_flashTP',
+        help='use flashTP accelerations for training',
         action='store_true',
     )
     ag.add_argument(
@@ -180,13 +196,14 @@ def add_parser(subparsers):
     cmd_parser_train(ag)
 
 
-def set_default_subparser(self, name, args=None, positional_args=0):
+def set_default_subparser(self, name):
     """default subparser selection. Call after setup, just before parse_args()
     name: is the name of the subparser to call by default
     args: if set is the argument list handed to parse_args()
 
     Hack copied from stack overflow
     """
+
     subparser_found = False
     for arg in sys.argv[1:]:
         if arg in ['-h', '--help']:  # global help if no subparser
@@ -199,13 +216,8 @@ def set_default_subparser(self, name, args=None, positional_args=0):
                 if sp_name in sys.argv[1:]:
                     subparser_found = True
         if not subparser_found:
-            # insert default in last position before global positional
-            # arguments, this implies no global options are specified after
-            # first positional argument
-            if args is None:
-                sys.argv.insert(len(sys.argv) - positional_args, name)
-            else:
-                args.insert(len(args) - positional_args, name)
+            # we don't have global option except -h. So simply put 'train' to 1
+            sys.argv.insert(1, name)
 
 
 argparse.ArgumentParser.set_default_subparser = set_default_subparser  # type: ignore
@@ -233,11 +245,18 @@ def main():
     ag.set_default_subparser('train')  # type: ignore
     args = ag.parse_args()
 
-    if args.command is None:  # backward compatibility
-        args.command = 'train'
-
     if args.command == 'train':
         run(args)
+    elif args.command in ['checkpoint', 'cp']:
+        checkpoint_cmd.run(args)
+    elif args.command in ['get_model', 'deploy']:
+        get_model_cmd.run(args)
+    elif args.command == 'graph_build':
+        graph_build_cmd.run(args)
+    elif args.command in ['inference', 'inf']:
+        inference_cmd.run(args)
+    elif args.command == 'patch_lammps':
+        patch_lammps_cmd.run(args)
     elif args.command == 'preset':
         preset_cmd.run(args)