+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 All notable changes to this project will be documented in this file.
 
 
+## [0.12.0-dev]
+### Added
+- Example config for fine-tuning the SevenNet-MF-ompa model
+- FlashTP support (https://github.com/SNU-ARC/flashTP)
+
+### Changed
+- ninja dependency added
+
+### Fixed
+- Multi modal model finetuning: Pass modal_map into config during multi-fidelity continual training #232
+- omat deploy parallel fixed
+
+
+## [0.11.2]
+### Changed
+- change cueq default config
+- assign cell volume 0 instead of eps for non-pbc system
+- remove redundant CELL and CELL_SHIFT
+
+### Fixed
+- Circular import in error_recorder
+
+
 ## [0.11.1]
 From here, the version of 'main' branch has 'devX' after it diverges from the latest stable version